Use a macro for ser/de implementations
