+++ src/tinystr16.rs
@@ -330,28 +330,4 @@ impl Into<u128> for TinyStr16 {
     }
 }
 
-#[cfg(feature = "serde")]
-impl serde::Serialize for TinyStr16 {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_str(self.as_str())
-    }
-}
-
-#[cfg(feature = "serde")]
-impl<'de> serde::Deserialize<'de> for TinyStr16 {
-    fn deserialize<D>(deserializer: D) -> Result<TinyStr16, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        use std::borrow::Cow;
-        use serde::de::Error as SerdeError;
-        use std::string::ToString;
-
-        let x: Cow<'de, str> = serde::Deserialize::deserialize(deserializer)?;
-        x.parse().map_err(|e: Error| SerdeError::custom(e.to_string()))
-    }
-}
-
+serde_impl!(TinyStr16, u128);
